+++ starlight-basic/src/main/java/org/atcraftmc/starlight/commands/WorldEditCommands.java
@@ -1,8 +1,21 @@
 package org.atcraftmc.starlight.commands;
 
+import com.sk89q.worldedit.EditSession;
+import com.sk89q.worldedit.MaxChangedBlocksException;
+import com.sk89q.worldedit.WorldEdit;
+import com.sk89q.worldedit.bukkit.BukkitAdapter;
+import com.sk89q.worldedit.bukkit.BukkitPlayer;
+import com.sk89q.worldedit.extension.input.InputParseException;
+import com.sk89q.worldedit.extension.input.ParserContext;
+import com.sk89q.worldedit.extension.platform.Actor;
+import com.sk89q.worldedit.function.pattern.Pattern;
+import com.sk89q.worldedit.math.BlockVector3;
+import com.sk89q.worldedit.util.formatting.text.TextComponent;
+import com.sk89q.worldedit.util.formatting.text.TranslatableComponent;
 import me.gb2022.commons.compatibility.APIIncompatibleException;
 import me.gb2022.gluon.module.ApplicationModule;
 import org.atcraftmc.qlib.command.BukkitCommand;
+import org.atcraftmc.qlib.command.assertion.CommandAssertionException;
 import org.atcraftmc.qlib.command.execute.CommandExecution;
 import org.atcraftmc.qlib.command.execute.CommandSuggestion;
 import org.atcraftmc.starlight.foundation.command.CommandProvider;
@@ -10,13 +23,124 @@ import org.atcraftmc.starlight.foundation.command.ModuleCommand;
 import org.atcraftmc.starlight.foundation.platform.Compatibility;
 import org.atcraftmc.starlight.framework.module.BukkitAbstractModule;
 import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
+import org.enginehub.piston.inject.InjectedValueAccess;
+import org.enginehub.piston.inject.Key;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.HashMap;
+import java.util.Optional;
 
 @ApplicationModule(id = "world-edit-commands")
-@CommandProvider({WorldEditCommands.MirrorCommand.class, WorldEditCommands.DrainWaterCommand.class, WorldEditCommands.FastBrushCommand.class})
+@CommandProvider({WorldEditCommands.MirrorCommand.class, WorldEditCommands.DrainWaterCommand.class, WorldEditCommands.FastBrushCommand.class, WorldEditCommands.BoxOutlineCommand.class})
 public final class WorldEditCommands extends BukkitAbstractModule {
     @Override
     public void checkCompatibility() throws APIIncompatibleException {
         Compatibility.requirePlugin("WorldEdit");
+    }
+
+    @BukkitCommand(name = "/outline-box", permission = "-worldedit.region.outline")
+    public static final class BoxOutlineCommand extends ModuleCommand<WorldEditCommands> {
+        public static Pattern parsePattern(Player player, String input) throws Exception {
+            BukkitPlayer wePlayer = BukkitAdapter.adapt(player);
+
+            ParserContext context = new ParserContext();
+            context.setActor(wePlayer);
+            context.setWorld(wePlayer.getWorld());
+
+            return WorldEdit.getInstance().getBlockFactory().parseFromInput(input, context);
+        }
+
+        public static void drawHollowBorder(Player player, Pattern pattern) throws Exception {
+
+            var wePlayer = BukkitAdapter.adapt(player);
+
+            var region = WorldEdit.getInstance().getSessionManager().get(wePlayer).getSelection(wePlayer.getWorld());
+            var change = 0;
+
+            try (EditSession editSession = WorldEdit.getInstance().newEditSession(wePlayer.getWorld())) {
+
+                var min = region.getMinimumPoint();
+                var max = region.getMaximumPoint();
+
+                var minX = min.getX();
+                var maxX = max.getX();
+                var minY = min.getY();
+                var maxY = max.getY();
+                var minZ = min.getZ();
+                var maxZ = max.getZ();
+
+                for (int x = minX; x <= maxX; x++) {
+                    set(editSession, x, minY, minZ, pattern);
+                    set(editSession, x, minY, maxZ, pattern);
+                    set(editSession, x, maxY, minZ, pattern);
+                    set(editSession, x, maxY, maxZ, pattern);
+                    change += 4;
+                }
+
+                for (int y = minY; y <= maxY; y++) {
+                    set(editSession, minX, y, minZ, pattern);
+                    set(editSession, minX, y, maxZ, pattern);
+                    set(editSession, maxX, y, minZ, pattern);
+                    set(editSession, maxX, y, maxZ, pattern);
+                    change += 4;
+                }
+
+                for (int z = minZ; z <= maxZ; z++) {
+                    set(editSession, minX, minY, z, pattern);
+                    set(editSession, minX, maxY, z, pattern);
+                    set(editSession, maxX, minY, z, pattern);
+                    set(editSession, maxX, maxY, z, pattern);
+                    change += 4;
+                }
+
+                editSession.flushSession();
+
+                wePlayer.printInfo(TranslatableComponent.of("worldedit.line.changed", TextComponent.of(change)));
+            }
+        }
+
+        static void set(EditSession editSession, int x, int y, double z, Pattern pattern) throws MaxChangedBlocksException {
+            editSession.setBlock(BlockVector3.at(x, y, z), pattern);
+        }
+
+        @Override
+        public void suggest(CommandSuggestion suggestion) {
+            var command = WorldEdit.getInstance().getPlatformManager().getPlatformCommandManager().getCommandManager();
+            var conv = command.getConverter(Key.of(Pattern.class));
+            var buf = suggestion.getBuffer();
+            var iva = new SimpleIVA();
+
+            iva.put(Actor.class, BukkitAdapter.adapt(suggestion.getSenderAsPlayer()));
+            var b = buf.get(buf.size() - 1);
+
+            conv.ifPresent((c) -> suggestion.suggest(0, c.getSuggestions(b, iva)));
+        }
+
+        @Override
+        public void execute(CommandExecution context) {
+            try {
+                var player = context.requireSenderAsPlayer();
+                var pattern = parsePattern(player, context.requireArgumentAt(0));
+
+                drawHollowBorder(player, pattern);
+            } catch (InputParseException e) {
+                context.getSender().sendMessage(ChatColor.RED + e.getMessage());
+            } catch (CommandAssertionException e) {
+                throw e;
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        private static final class SimpleIVA extends HashMap<Class<?>, Object> implements InjectedValueAccess {
+
+            @Override
+            public <T> @NotNull Optional<T> injectedValue(@NotNull Key<T> key, @NotNull InjectedValueAccess injectedValueAccess) {
+                return (Optional<T>) Optional.ofNullable(this.get(key.getTypeToken().getRawType()));
+            }
+        }
     }
 
     @BukkitCommand(name = "/mirror")