+++ starlight-extensions/starlight-ext-lobby/src/main/java/org/atcraftmc/quark/lobby/DefaultInventory.java
@@ -1,13 +1,17 @@
 package org.atcraftmc.quark.lobby;
 
-import me.gb2022.commons.nbt.NBTTagCompound;
 import me.gb2022.commons.reflect.AutoRegister;
 import me.gb2022.gluon.Registrations;
+import me.gb2022.gluon.module.ApplicationModule;
+import org.atcraftmc.qlib.command.BukkitCommand;
+import org.atcraftmc.qlib.command.execute.CommandExecution;
+import org.atcraftmc.qlib.command.execute.CommandSuggestion;
+import org.atcraftmc.starlight.framework.module.SLCommandModule;
+import org.atcraftmc.starlight.migration.MessageAccessor;
+import org.atcraftmc.starlight.shared.Configurations;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
 import org.bukkit.entity.HumanEntity;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryCloseEvent;
@@ -17,16 +21,9 @@ import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
-import org.atcraftmc.starlight.core.data.ModuleDataService;
-import org.atcraftmc.qlib.command.BukkitCommand;
-import org.atcraftmc.starlight.framework.module.SLCommandModule;
-import me.gb2022.gluon.module.ApplicationModule;
-import org.atcraftmc.starlight.migration.MessageAccessor;
 
-import java.util.List;
-
-@ApplicationModule(version = "1.0")
-@BukkitCommand(name = "default-inventory", playerOnly = true)
+@ApplicationModule(id = "lobby-default-inventory")
+@BukkitCommand(name = "lobby-default-inventory", playerOnly = true)
 @AutoRegister(Registrations.SERVER_EVENT)
 public final class DefaultInventory extends SLCommandModule {
     public static final String TITLE = ChatColor.LIGHT_PURPLE + "Default Inventory Editor";
@@ -68,24 +65,14 @@ public final class DefaultInventory extends SLCommandModule {
     }
 
     @Override
-    public void enable() {
+    public void enable() throws Exception {
         super.enable();
-
-        NBTTagCompound tag = ModuleDataService.getEntry(this.getFullId()).getCompoundTag("inventory");
-
-        if (tag == null) {
-            return;
-        }
-
-        for (String s : tag.getTagMap().keySet()) {
-            inventory.setItem(Integer.parseInt(s), ItemStack.deserializeBytes(tag.getByteArray(s)));
-        }
+        load();
     }
 
     @Override
-    public void disable() {
+    public void disable() throws Exception {
         super.disable();
-        this.save();
     }
 
     @EventHandler
@@ -96,12 +83,28 @@ public final class DefaultInventory extends SLCommandModule {
 
         this.save();
 
-        MessageAccessor.send(this.language, event.getPlayer(), "edit-complete");
+        MessageAccessor.send(this.language(), event.getPlayer(), "edit-complete");
+    }
+
+    private void load() {
+        var conf = Configurations.standalone("lobby-default-inventory");
+
+        this.inventory.clear();
+
+        for (var k : conf.getKeys(false)) {
+            var i = Integer.parseInt(k);
+
+            this.inventory.setItem(i, conf.getItemStack(k));
+        }
     }
 
     private void save() {
-        int index = 0;
-        NBTTagCompound entry = new NBTTagCompound();
+        var conf = Configurations.standalone("lobby-default-inventory");
+        var index = 0;
+
+        for (var k : conf.getKeys(false)) {
+            conf.set(k, null);
+        }
 
         for (ItemStack stack : this.inventory.getContents()) {
             if (stack == null) {
@@ -109,34 +112,35 @@ public final class DefaultInventory extends SLCommandModule {
                 continue;
             }
 
-            entry.setByteArray(String.valueOf(index), stack.serializeAsBytes());
+            conf.set(String.valueOf(index), stack);
             index++;
         }
 
-        NBTTagCompound tag = ModuleDataService.getEntry(this.getFullId());
-        tag.setCompoundTag("inventory", entry);
-        ModuleDataService.save(this.getFullId());
+        Configurations.saveStandalone(conf);
     }
 
     @Override
-    public void onCommand(CommandSender sender, String[] args) {
-        switch (args[0]) {
+    public void suggest(CommandSuggestion suggestion) {
+        suggestion.suggest(0, "cover", "edit", "reload");
+    }
+
+    @Override
+    public void execute(CommandExecution context) {
+        var sender = context.getSender();
+
+        switch (context.requireEnum(0, "cover", "edit", "reload")) {
             case "edit" -> {
-                ((Player) sender).openInventory(this.inventory);
-                MessageAccessor.send(this.language, sender, "edit-start");
+                context.requireSenderAsPlayer().openInventory(this.inventory);
+                MessageAccessor.send(this.language(), sender, "edit-start");
             }
             case "cover" -> {
-                cover((Player) sender);
-                MessageAccessor.send(this.language, sender, "cover");
+                cover(context.requireSenderAsPlayer());
+                MessageAccessor.send(this.language(), sender, "cover");
             }
-        }
-    }
-
-    @Override
-    public void onCommandTab(CommandSender sender, String[] buffer, List<String> tabList) {
-        if (buffer.length == 1) {
-            tabList.add("cover");
-            tabList.add("edit");
+            case "reload" -> {
+                load();
+                MessageAccessor.send(this.language(), sender, "reload");
+            }
         }
     }
 }