+++ starlight-extensions/starlight-ext-lobby/src/main/java/org/atcraftmc/quark/lobby/LobbyParkour.java
@@ -1,4 +1,0 @@
-package org.atcraftmc.quark.lobby;
-
-public class LobbyParkour {
-}