+++ starlight-extensions/starlight-ext-lobby/src/main/java/org/atcraftmc/quark/lobby/PlayerProtect.java
@@ -2,16 +2,16 @@ package org.atcraftmc.quark.lobby;
 
 import me.gb2022.commons.reflect.AutoRegister;
 import me.gb2022.gluon.Registrations;
+import me.gb2022.gluon.module.ApplicationModule;
 import org.atcraftmc.starlight.framework.module.BukkitAbstractModule;
+import org.atcraftmc.starlight.migration.ConfigAccessor;
 import org.bukkit.entity.EntityType;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.player.PlayerMoveEvent;
-import me.gb2022.gluon.module.ApplicationModule;
-import org.atcraftmc.starlight.migration.ConfigAccessor;
 
 @AutoRegister(Registrations.SERVER_EVENT)
-@ApplicationModule(version = "1.0.3")
+@ApplicationModule(version = "1.0.3", id = "lobby-player-protect")
 public final class PlayerProtect extends BukkitAbstractModule {
     @EventHandler
     public void onPlayerDamage(EntityDamageEvent event) {
@@ -26,6 +26,6 @@ public final class PlayerProtect extends BukkitAbstractModule {
         if (event.getFrom().getY() > ConfigAccessor.getInt(this.config(), "lowest-y")) {
             return;
         }
-        event.getPlayer().teleport(event.getPlayer().getWorld().getSpawnLocation().add(0.5,0.5,0.5));
+        event.getPlayer().teleport(event.getPlayer().getWorld().getSpawnLocation().add(0.5, 0.5, 0.5));
     }
 }